Apply patches

+++ SAX2.c
@@ -1908,7 +1908,7 @@ skip:
 	    else {
 	        ret->line = 65535;
 		if (ctxt->options & XML_PARSE_BIG_LINES)
-		    ret->psvi = (void *) (long) ctxt->input->line;
+		    ret->psvi = (void *) (intptr_t) ctxt->input->line;
 	    }
 	}
     }
